[Seq] Moved SeqToSV to Conversions
