+++ lib/CAPI/CMakeLists.txt
@@ -1,2 +1,3 @@
+add_subdirectory(Conversion)
 add_subdirectory(ExportVerilog)
 add_subdirectory(Dialect)
