+++ lib/Conversion/CMakeLists.txt
@@ -7,6 +7,7 @@ add_subdirectory(HandshakeToFIRRTL)
 add_subdirectory(HandshakeToHW)
 add_subdirectory(LLHDToLLVM)
 add_subdirectory(MooreToCore)
+add_subdirectory(SeqToSV)
 add_subdirectory(SCFToCalyx)
 add_subdirectory(StandardToHandshake)
 add_subdirectory(StandardToStaticLogic)
