+++ lib/Conversion/PassDetail.h
@@ -80,6 +80,10 @@ namespace staticlogic {
 class StaticLogicDialect;
 } // namespace staticlogic
 
+namespace seq {
+class SeqDialect;
+} // namespace seq
+
 namespace sv {
 class SVDialect;
 } // namespace sv
