+++ lib/Dialect/Seq/CMakeLists.txt
@@ -11,8 +11,8 @@
 
 add_circt_dialect_library(CIRCTSeq
   SeqDialect.cpp
-  SeqOps.cpp
   SeqPasses.cpp
+  SeqOps.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${CIRCT_MAIN_INCLUDE_DIR}/circt/Dialect/Seq
