+++ lib/Dialect/Seq/SeqPasses.cpp
@@ -16,87 +16,3 @@
 using namespace mlir;
 using namespace circt;
 using namespace seq;
-
-namespace circt {
-namespace seq {
-#define GEN_PASS_CLASSES
-#include "circt/Dialect/Seq/SeqPasses.h.inc"
-} // namespace seq
-} // namespace circt
-
-namespace {
-struct SeqToSVPass : public LowerSeqToSVBase<SeqToSVPass> {
-  void runOnOperation() override;
-};
-} // anonymous namespace
-
-namespace {
-/// Lower CompRegOp to `sv.reg` and `sv.alwaysff`. Use a posedge clock and
-/// synchronous reset.
-struct CompRegLower : public OpConversionPattern<CompRegOp> {
-public:
-  using OpConversionPattern::OpConversionPattern;
-
-  LogicalResult
-  matchAndRewrite(CompRegOp reg, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const final {
-    Location loc = reg.getLoc();
-
-    auto svReg = rewriter.create<sv::RegOp>(loc, reg.getResult().getType(),
-                                            reg.nameAttr());
-    svReg->setDialectAttrs(reg->getDialectAttrs());
-
-    // If the seq::CompRegOp has an inner_sym attribute, set this for the
-    // sv::RegOp inner_sym attribute.
-    if (reg.sym_name().hasValue())
-      svReg.inner_symAttr(reg.sym_nameAttr());
-
-    auto regVal = rewriter.create<sv::ReadInOutOp>(loc, svReg);
-    if (reg.reset() && reg.resetValue()) {
-      rewriter.create<sv::AlwaysFFOp>(
-          loc, sv::EventControl::AtPosEdge, reg.clk(), ResetType::SyncReset,
-          sv::EventControl::AtPosEdge, reg.reset(),
-          [&]() { rewriter.create<sv::PAssignOp>(loc, svReg, reg.input()); },
-          [&]() {
-            rewriter.create<sv::PAssignOp>(loc, svReg, reg.resetValue());
-          });
-    } else {
-      rewriter.create<sv::AlwaysFFOp>(
-          loc, sv::EventControl::AtPosEdge, reg.clk(),
-          [&]() { rewriter.create<sv::PAssignOp>(loc, svReg, reg.input()); });
-    }
-
-    rewriter.replaceOp(reg, {regVal});
-    return success();
-  }
-};
-
-} // namespace
-void SeqToSVPass::runOnOperation() {
-  ModuleOp top = getOperation();
-  MLIRContext &ctxt = getContext();
-
-  ConversionTarget target(ctxt);
-  target.addIllegalDialect<SeqDialect>();
-  target.addLegalDialect<sv::SVDialect>();
-  RewritePatternSet patterns(&ctxt);
-  patterns.add<CompRegLower>(&ctxt);
-
-  if (failed(applyPartialConversion(top, target, std::move(patterns))))
-    signalPassFailure();
-}
-
-namespace circt {
-namespace seq {
-std::unique_ptr<OperationPass<ModuleOp>> createSeqLowerToSVPass() {
-  return std::make_unique<SeqToSVPass>();
-}
-} // namespace seq
-} // namespace circt
-
-namespace {
-#define GEN_PASS_REGISTRATION
-#include "circt/Dialect/Seq/SeqPasses.h.inc"
-} // namespace
-
-void circt::seq::registerSeqPasses() { registerPasses(); }
