+++ tools/firtool/CMakeLists.txt
@@ -13,6 +13,7 @@ target_link_libraries(firtool PRIVATE
   CIRCTFIRRTLTransforms
   CIRCTHWTransforms
   CIRCTSVTransforms
+  CIRCTSeqToSV
 
   MLIRParser
   MLIRSupport
